[CD-210] Refactor: move create_dataset to utils file
